% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{age_from_ages}
\alias{age_from_ages}
\title{Calculate age in years from ages in other units}
\usage{
age_from_ages(age_y, age_m, age_d)
}
\description{
Calculate age in years from ages in other units
}
\keyword{internal}
