% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{age_from_dates}
\alias{age_from_dates}
\title{age from birth and reference dates}
\usage{
age_from_dates(birth.date, ref.date = Sys.Date(), unit = c("year", "day"))
}
\arguments{
\item{birth.date}{Date of birth, either as a \code{Date} or something which
will be converted to a \code{Date}}

\item{ref.date}{Date at which to calculate age, defaults to current date,
either as a \code{Date} or something which will be converted to a
\code{Date}}

\item{unit}{character of length, one of "year" or "day".}
}
\value{
integer vector
}
\description{
Calculate age at time of reference date, based on birth date.
}
\examples{
age_from_dates("2014-11-08", "2014-12-31", unit = "day")
age_from_dates("1981-07-09", "2014-06-29", unit = "year")
}
