% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_m}
\alias{age_m}
\alias{age_d}
\title{Calculate age in years from other units}
\usage{
age_m(age_m)

age_d(age_d)
}
\arguments{
\item{age_m}{Months}

\item{age_d}{Days}
}
\description{
Calculate age in years from other units
}
\examples{
age_m(12)
age_m(1)
}
\keyword{manip}
