% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumonary.R
\name{alveolar_paO2_mmHg}
\alias{alveolar_paO2_mmHg}
\title{alveolar gas equation}
\usage{
alveolar_paO2_mmHg(FiO2 = 0.209, RQ = 0.8, paCO2_mmHg = 40,
  Patm_mmHg = 760, paH2O_mmHg = 47)
}
\arguments{
\item{FiO2}{fraction of oxygen in inspired gas, from 0 to 1, default is (dry)
room air}

\item{RQ}{respiratory quotient, i.e. the ratio of CO2 produced to oxygen
consumed, usually between around 0.7 and 1.0, but can legitimately be
greater than 1.0. Default it 0.8.}

\item{paCO2}{partial pressure of CO2 in alveolus, which can be approximated
as the end-tidal pCO2}

\item{paH2O}{partial pressure of water vapor, defaults to 6.25 kPa (47 mmHg)
which is appropriate for body temperature}

\item{Patm}{atmospheric pressure in kPa}
}
\description{
Estimate paO2 in alveolus based on atmospheric pressure, FiO2, PaCO2, and
respiratory quotient
}
\examples{
# vary RQ
rq <- seq(0.6, 1.4, 0.05)
plot(rq, alveolar_paO2_mmHg(RQ = rq))

# 100\% FiO2 at typical atmospheric pressure
alveolar_paO2_mmHg(FiO2 = 1)

# hyperbaric oxygen at 100\%, 2 atmospheres
alveolar_paO2_mmHg(FiO2 = 1, Patm_mmHg = 1520)
}

