% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{atm_pres}
\alias{atm_pres}
\alias{atm_pres_frac}
\title{Get mean atmospheric pressure at given altitude}
\usage{
atm_pres(altitude_m)

atm_pres_frac(altitude_m)
}
\arguments{
\item{altitude_m}{Altitude above mean sea level in meters}
}
\value{
Pressure in pascals
}
\description{
Get mean atmospheric pressure at given altitude
}
\section{Functions}{
\itemize{
\item \code{atm_pres_frac}: Get fraction of mean atomspheric pressure at sea level
}}

\examples{
atm_pres(-430.5) # Dead Sea
atm_pres(0)
atm_pres(3440) # Namche Bazaar
atm_pres(4260) # Dingboche
atm_pres(5364) # Everest Base Camp
atm_pres(6000) # Camp 1
atm_pres(6400) # Camp 2
atm_pres(7200) # Camp 3
atm_pres(7950) # Camp 4
atm_pres(8850) # Everest summit
atm_pres_frac(8850) # fraction of sea level pressure on Everest
}
