% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi_adult}
\alias{bmi_adult}
\alias{bmi_adult_ins_lbs}
\title{Body Mass Index (BMI) for adults}
\usage{
bmi_adult(heightm, weightkg, ...)

bmi_adult_ins_lbs(heightin, weightlb, ...)
}
\arguments{
\item{heightm}{single numeric, height in meters}

\item{weightkg}{numeric vector of weight(s) in kg}

\item{...}{passed to validation}

\item{heightin}{height in inches}

\item{weightlb}{weight in pounds}
}
\description{
Calculate body mass index using weight in kg / (height in metres ^ 2)
}
\examples{
bmi_adult(1.6, 120)
bmi_adult(2, 75)
bmi_adult_ins_lbs(72, 200)
}

