% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bsa_adult}
\alias{bsa_adult}
\title{Estimate body surface area of an adult}
\usage{
bsa_adult(height_m, weight_kg, ...)
}
\arguments{
\item{height_m}{single numeric, height in meters}

\item{weight_kg}{numeric vector of weight(s) in kg}

\item{...}{passed to validation}
}
\value{
numeric vector
}
\description{
\code{bsa_adult} Estimate body surface area of an
  adult using \code{sqrt(wt*ht)/6} TODO: reference for this.
}
\examples{
bsa_adult(2, 80)
bsa_adult(1.5, 80)
}
