% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{french_to_diameter_mm}
\alias{french_to_diameter_mm}
\alias{diameter_mm_to_french}
\title{French to diameter}
\usage{
french_to_diameter_mm(x)

diameter_mm_to_french(x)
}
\arguments{
\item{x}{Size in French units, or mm}
}
\description{
Convert French size of a catheter to diameter in mm. Currently accepts or
returns non-integer French values
}
\concept{unit conversions}
