% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStandardTemperature}
\alias{getStandardTemperature}
\title{get standard temperature at a given altitude}
\usage{
getStandardTemperature(altitude_km)
}
\arguments{
\item{altitude_km}{Altitude in kilometers}
}
\description{
altitude_km = earth_radius * altitude / (earth_radius + altitude) All in
km Temperature is in kelvins = 273.15 + Celsius
}
\references{
Below 51 km: Practical Meteorology by Roland Stull, pg 12. Above 51 km:
http://www.braeunig.us/space/atmmodel.htm Validation data:
https://www.avs.org/AVS/files/c7/c7edaedb-95b2-438f-adfb-36de54f87b9e.pdf
}
\keyword{internal}
