% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acidbase.R
\name{henderson_hasselbach}
\alias{henderson_hasselbach}
\title{pH by Henderson Hasselbach equation}
\usage{
henderson_hasselbach(bicarbonate, ppCO2)
}
\arguments{
\item{bicarbonate}{mmol/L}

\item{ppCO2}{partial pressure of carbon dioxide in mmHg}
}
\description{
Calculate the pH based on bicarbonate and partial pressure of CO2
}
\examples{
 bicarbonate <- seq(10, 50, 5)
 ppCO2 <- seq(20, 70, 10)
 bc <- rep(bicarbonate, length(ppCO2))
 pp <- rep(ppCO2, each = length(bicarbonate))
 acidbase <- matrix(henderson_hasselbach(bc, pp), nrow = 9, ncol = 6)
 rownames(acidbase) <- paste("bicarb", bicarbonate)
 colnames(acidbase) <- paste("PaCO2", ppCO2)
 acidbase
}

