% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{ideal_or_actual_weight}
\alias{ideal_or_actual_weight}
\title{Get ideal weight if possible}
\usage{
ideal_or_actual_weight(male = NULL, height_m = NULL, weight_kg = NULL,
  age_y = NULL)
}
\description{
Get ideal weight if possible
}
\keyword{internal}
