% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{is_adult}
\alias{is_adult}
\title{Is age >= 18 years}
\usage{
is_adult(age_y)
}
\arguments{
\item{age_y}{Numeric vector of age in years. Bear in mind that age is not
handled with extreme precision, since it is only used for approximating
physiologic characteristics.}
}
\description{
Is age >= 18 years
}
