% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\docType{data}
\name{Pa_to_torr}
\alias{Pa_to_torr}
\title{Conversion factor from Pa to torr (mmHg)}
\format{An object of class \code{numeric} of length 1.}
\usage{
Pa_to_torr
}
\description{
The conversion is exactly 760 / 101325
}
\seealso{
Other physics: \code{\link{pres_atm_kPa}},
  \code{\link{svp_sea_level}}, \code{\link{temp_c_to_k}}
}
\concept{physics}
\keyword{datasets}
