% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peds_anes_calc.R
\name{peds_anes_calc}
\alias{peds_anes_calc}
\title{Launch pedatric anesthesia calculator shiny application}
\usage{
peds_anes_calc(...)
}
\description{
Launch pedatric anesthesia calculator shiny application
}
