% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{svp_sea_level}
\alias{svp_sea_level}
\title{Saturation vapor pressure of water at sea level}
\usage{
svp_sea_level(temp_k)
}
\arguments{
\item{temp_k}{Temperature in Kelvin}
}
\description{
Saturation vapor pressure of water at sea level
}
