% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{temp_c_to_k}
\alias{temp_c_to_k}
\title{Temperature in Kelvin from Celsius}
\usage{
temp_c_to_k(temp_c)
}
\arguments{
\item{temp_c}{Temperature in Celsius}
}
\description{
Temperature in Kelvin from Celsius
}
