% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{adj_weight_adult}
\alias{adj_weight_adult}
\title{adjusted body weight}
\usage{
adj_weight_adult(height_m, weight_kg, male, ...)
}
\arguments{
\item{height_m}{single numeric, height in meters}

\item{weight_kg}{weight in kg, may be a vector}

\item{male}{logical value(s) whether patient is male. TRUE or FALSE.}

\item{...}{passed to validation}
}
\description{
returns ideal weight + 40% of difference between ideal and
  actual weights. Ideal weight is calculated using default algorithm. TODO:
  is downward adjustment valid?
}
\examples{
adj_weight_adult(1.6, 120, male = FALSE)
}
