% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age_m_to_y}
\alias{age_m_to_y}
\alias{age_d_to_y}
\alias{age_d_to_m}
\title{Calculate age in years from other units}
\usage{
age_m_to_y(age_m)

age_d_to_y(age_d)

age_d_to_m(age_d)
}
\arguments{
\item{age_m}{Months}

\item{age_d}{Days}
}
\description{
Calculate age in years from other units
}
\examples{
age_m_to_y(12)
age_m_to_y(1)
}
\keyword{manip}
