% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{as_age_y}
\alias{as_age_y}
\title{Calculate age in years from ages in other units}
\usage{
as_age_y(age_y = 0, age_m = 0, age_d = 0, birth_date = NULL)
}
\description{
Only one of years, months or days should be supplied. If birth date is known,
we can use a calendar to do this exactly. If not known, then we let lubridate
apporximate.
}
\details{
\code{lubridate::duration} gives mathematically predictable results,
\code{lubridate::period} depends on calendar. So, giving a DoB will use the
latter.
}
\keyword{internal}
