% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocs.R
\docType{data}
\name{deadspace_things_ml}
\alias{deadspace_things_ml}
\title{Apparatus dead space for common things}
\format{An object of class \code{list} of length 5.}
\usage{
deadspace_things_ml
}
\description{
Internal volumes reported in milliliters, based on Feldman, 2015.
}
\references{
Feldman, J, 2015, Pediatric Anesthesia
}
\seealso{
Other airway equipment: \code{\link{deadspace_total}},
  \code{\link{ett_size_by_age}}, \code{\link{ett_vol_ml}}
}
\concept{airway equipment}
\concept{dead-space}
\concept{deadspace}
\keyword{datasets}
\keyword{internal}
\keyword{list}
