% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventilation.R
\name{deadspace_total}
\alias{deadspace_total}
\alias{deadspace_anatomic}
\title{Estimate ventilatory dead space}
\usage{
deadspace_total(height_m, age_y = NULL, gender = NULL, elbow_ml = 10,
  humidifier_ml = 7, ett_diameter_mm = NULL)

deadspace_anatomic(height_m = NULL, weight_kg = NULL, age_y = NULL)
}
\arguments{
\item{height_m}{Height in m}

\item{age_y}{Age in years, optional for estimating ETT and HME sizes
automatically}

\item{elbow_ml}{Numeric: volume of elbow of breathing circuit in ml}

\item{humidifier_ml}{Numeric: volume of humidifier of breathing circuit in ml}

\item{ett_diameter_mm}{Numeric: internal diameter of endotrachel tube.
Default is \code{NULL} which would estimate this from the age of patient}

\item{gender_f}{Logical: female if \code{TRUE}. Default is not to specify
gender.}
}
\value{
estimate of anatomic deadspace in ml
}
\description{
Estimate ventilatory dead space
}
\section{Functions}{
\itemize{
\item \code{deadspace_anatomic}: Estimate anatomic deadspace
}}

\references{
http://www.atsjournals.org/doi/abs/10.1164/arrd.1971.104.2.215
  http://rc.rcjournal.com/content/53/7/885.short
  https://www.ncbi.nlm.nih.gov/pubmed/8727530
}
