% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getStandardTemperature}
\alias{getStandardTemperature}
\title{get standard temperature at a given altitude}
\usage{
getStandardTemperature(altitude_km)
}
\arguments{
\item{altitude_km}{Altitude in kilometers}
}
\description{
\code{altitude_km = earth_radius * altitude / (earth_radius + altitude)} All in km Temperature is in Kelvin.
}
\seealso{
\code{\link{pres_atm_kPa}}
}
\keyword{internal}
