% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\docType{data}
\name{med_conv}
\alias{med_conv}
\title{Conversion factors and reference ranges for lab values}
\format{data frame in a tidy format, with the following columns, with example
 values:
 - Analyte _Carbon dioxide, _Pco2_
 - Specimen _Arterial blood_
 - Reference Range, Conventional Unit _35-45_
 - Conventional Unit _mm Hg_
 - Conversion Factor (Multiply by) _0.133_
 - Reference Range, SI Unit _47.-5.9_
 - SI Unit _kPa_
 @source \url{http://www.amamanualofstyle.com/page/si-conversion-calculator}}
\usage{
med_conv
}
\description{
A few hundred lab values with _Conventional_ (US) and SI units, with a
scaling factor and reference ranges.
}
\keyword{datasets}
