% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiology-package.R
\docType{package}
\name{physiology-package}
\alias{physiology-package}
\alias{_PACKAGE}
\alias{physiology}
\title{physiology}
\description{
A variety of formulae are provided for estimation
    of physiologic characteristics of infants, children, and adults.
    Calculations include: body surface area, ideal weight, airway
    dead-space, the alveolar gas equation, and GFR.  Each formula is
    referenced to the original publication. Future functions will cover
    more material with a focus on anaesthesia, critical care and
    peri-operative medicine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jackwasey.github.io/physiology/}
  \item Report bugs at \url{https://github.com/jackwasey/physiology/issues}
}

}
\author{
\strong{Maintainer}: Jack O. Wasey \email{jack@jackwasey.com} (0000-0003-3738-4637) [copyright holder]

Other contributors:
\itemize{
  \item Bill Denney \email{wdenney@humanpredictions.com} (0000-0002-5759-428X) [contributor]
}

}
\concept{blood}
\concept{ideal weight}
\concept{medicine}
\concept{physiology}
\concept{renal}
\keyword{misc}
\keyword{utilities}
