% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{svp_sea_level}
\alias{svp_sea_level}
\title{Saturation vapor pressure of water at sea level}
\usage{
svp_sea_level(temp_k)
}
\arguments{
\item{temp_k}{Temperature in Kelvin}
}
\description{
Saturation vapor pressure of water at sea level
}
\seealso{
Other physics: \code{\link{Pa_to_torr}},
  \code{\link{pres_atm_kPa}}, \code{\link{temp_c_to_k}}
}
\concept{physics}
