% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{temp_c_to_k}
\alias{temp_c_to_k}
\title{Temperature in Kelvin from Celsius}
\usage{
temp_c_to_k(temp_c)
}
\arguments{
\item{temp_c}{Temperature in Celsius}
}
\description{
Temperature in Kelvin from Celsius
}
\seealso{
Other physics: \code{\link{Pa_to_torr}},
  \code{\link{pres_atm_kPa}}, \code{\link{svp_sea_level}}
}
\concept{physics}
