% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_height}
\alias{valid_height}
\alias{valid_height_adult}
\alias{valid_weight}
\alias{valid_weight_adult}
\alias{valid_age}
\alias{valid_age_adult}
\title{validate phsyiologic input parameters}
\usage{
valid_height(height_m, ht.min = 0.1, ht.max = 2.5, ht.min.hard = 0.001,
  ht.max.hard = 3, extramsg = "", do.warn = TRUE, do.stop = FALSE)

valid_height_adult(height_m, ht.min = 0.5, ht.max = 2.5,
  ht.min.hard = 0.001, ht.max.hard = 3, extramsg = "", do.warn = TRUE,
  do.stop = FALSE)

valid_weight(weight_kg, wt.min = 0.1, wt.max = 300, wt.min.hard = 0,
  wt.max.hard = 600, extramsg = "", do.warn = TRUE, do.stop = FALSE)

valid_weight_adult(weight_kg, wt.min = 5, wt.max = 300, wt.min.hard = 0,
  wt.max.hard = 600, extramsg = "", do.warn = TRUE, do.stop = FALSE)

valid_age(age, age.min = 0, age.max = 150, age.min.hard = 1e-05,
  age.max.hard = 150, extramsg = "", do.warn = TRUE, do.stop = FALSE)

valid_age_adult(age.years, age.min = 18, age.max = 150, age.min.hard = 17,
  age.max.hard = 150, extramsg = "", do.warn = TRUE, do.stop = FALSE)
}
\arguments{
\item{height_m}{single numeric, height in meters}

\item{ht.min}{minimum height below which to warn if \code{warn = TRUE}}

\item{ht.max}{maximum height above which to warn if \code{warn = TRUE}}

\item{ht.min.hard}{minimum height below which to warn regardless of
\code{warn}}

\item{ht.max.hard}{maximum height above which to warn if \code{warn}}

\item{extramsg}{single character string with additional message to append,
default is ""}

\item{do.warn}{single logical, if TRUE, will give warnings outside of soft
limits}

\item{do.stop}{single logical, \code{stop} instead of warning if any values
outside hard limits}

\item{weight_kg}{numeric vector of weight(s) in kg}

\item{wt.min}{minimum height below which to warn if \code{warn = TRUE}}

\item{wt.max}{maximum height above which to warn if \code{warn = TRUE}}

\item{wt.min.hard}{minimum height below which to warn regardless of \code{warn}}

\item{wt.max.hard}{maximum height above which to warn if \code{warn}}

\item{age}{numeric of years or \code{lubridate} Period, i.e. a fixed length
of time, which does not vary. If birth dates of individuals are known, the
user should use lubridate to convert Durations from birthdates to Periods.}

\item{age.min}{minimum age below which to warn if \code{warn = TRUE}}

\item{age.max}{maximum age above which to warn if \code{warn = TRUE}}

\item{age.min.hard}{minimum age below which to warn regardless of
\code{warn}}

\item{age.max.hard}{maximum age above which to warn if \code{warn}}
}
\description{
User may generate warnings for unreasonable or obviously
  erroneous heights. TODO: use age to validate further
}
